requirements.txt
aniso8601==9.0.1
astroid==2.9.0
attrs==21.2.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
dnspython==2.1.0
Flask-RESTful==0.3.9
Flask==2.0.2
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.3
platformdirs==2.4.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
PyJWT==2.3.0
pylint==2.12.2
pymongo==4.0
pyparsing==3.0.6
pytest==6.2.5
pytz==2021.3
requests==2.26.0
six==1.16.0
toml==0.10.2
typing-extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
